requirements.txt
pypdf~=6.6.0
chromadb~=1.4.0
openai~=2.14.0
python-dotenv~=1.2.1
pathlib~=1.0.1
fitz~=0.0.1.dev2
pillow~=12.1.0
torch~=2.9.1
transformers~=4.57.5